Extract more utils
