+++ addon/index.js
@@ -1,0 +1,7 @@
+import { getInstance } from './components/ember-islands';
+
+export function reconcile() {
+  getInstance().reconcile();
+}
+
+export { getInstance as pleaseDontUseThisExportToGetTheEmberIslandsInstance };
