+++ addon/utils/reconciler.js
@@ -2,20 +2,13 @@ import Ember from 'ember';
 const { guidFor } = Ember;
 
 export default class Reconciler {
-  constructor() {
+  constructor(callbacks) {
     this.renderedComponents = {};
-  }
-
-  addRenderedComponent({ instance, container }) {
-    let id = guidFor(container);
-    this.renderedComponents[id] = instance;
+    this.callbacks = callbacks;
   }
 
   reconcileAgainst(componentsToRender) {
     let nextRenderedComponents = {};
-    let componentsToUpdate = [];
-    let componentsToInitialRender = [];
-    let componentsToDestroy = [];
 
     componentsToRender.forEach(c => {
       let elementId = guidFor(c.element);
@@ -23,26 +16,21 @@ export default class Reconciler {
       let instance = this.renderedComponents[elementId];
       if (instance) {
         c.instance = instance;
-        componentsToUpdate.push(c);
-        nextRenderedComponents[elementId] = instance;
+        this.callbacks.update(c);
       } else {
-        componentsToInitialRender.push(c);
+        instance = this.callbacks.initialRender(c);
       }
+
+      nextRenderedComponents[elementId] = instance;
     });
 
     for (let key in this.renderedComponents) {
-      if (!nextRenderedComponents[key]) {
-        componentsToDestroy.push(this.renderedComponents[key]);
+      if (!(key in nextRenderedComponents)) {
+        this.callbacks.destroy(this.renderedComponents[key]);
       }
     }
 
     this.renderedComponents = nextRenderedComponents;
-
-    return {
-      initialRender: componentsToInitialRender,
-      update: componentsToUpdate,
-      destroy: componentsToDestroy
-    };
   }
 
   renderedComponentsAsArray() {
@@ -60,5 +48,3 @@ export default class Reconciler {
     }
   }
 }
-
-// EIComponent is probably a thing
