+++ app/components/ember-islands.js
@@ -1,1 +1,1 @@
-export { default, getInstance } from 'ember-islands/components/ember-islands';
+export { default } from 'ember-islands/components/ember-islands';
